third_party/WebKit/LayoutTests/http/tests/websocket/tests/hybi/extensions-with-experimental-flag.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test WebSocket.extensions attribute.");

window.jsTestIsAsync = true;
window.internals.settings.setExperimentalWebSocketEnabled(true);

var ws = new WebSocket("ws://localhost:8880/websocket/tests/hybi/echo-requested-extensions");
debug("Check the value before opening the connection:");
shouldBeEqualToString("ws.extensions", "");

// extensions attribute is read-only.
debug("Check if the value is read only:");
shouldBeEqualToString("ws.extensions = 'foo'; ws.extensions", "");

ws.onopen = function(event)
{
    debug("Check the value after the connection is established:");
    shouldBeEqualToString("ws.extensions", "permessage-deflate");
    ws.send("hi!");
};

ws.onmessage = function(event)
{
    debug("Check the value of the Sec-WebSocket-Extensions header sent to the server:");
    shouldBeEqualToString("event.data", "permessage-deflate; client_max_window_bits, x-webkit-deflate-frame");
    finishJSTest();
};

</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
